The methods showFromCreateThread and saveThread have been implemented.
The methods showFromCreateThread and saveThread have been implemented. The subName is used by the thread object with the a temp userName of "john". The model containing the newly created thread object is passed to the thread-form.jsp. Once the user fills out the form. The page will be redirected to the showSub function in the SubController.