Checks if the user is the creator of the comment for editing.
Added in HttpServletResponse and request. The request is to check the currently logged in user's name. If the user name matches the comment's creator name. That user can edit the comment. If not, the response sendError method will be called. It will be set to HttpServletResponse.SC_FORBIDDEN with the message "Forbidden access".