Fix editing deleted comments.
If the comment has been marked deleted the user can't edit it. First place the form and the link for edit comment inside the if statement which checks if the comment is deleted.

Inside the comment controller saveComment check if the comment is already deleted and if its already deleted throw a 404 forbidden access.